Break `ReadOnlyBatchSpecForm` into its own component, add separate alert body if execution finished
